Makes Throwable to keep the backing js error.
Earlier I did some refactorings towards making GWT exceptions
look like real Javascript Errors. One of them was to make
StackTraceCreator behave like Chrome's Error.captureStackTrace
so that it will store a stack property to the passed object.
I was going to follow up with changing Throwable's prototype
to Error.prototype so that it would be a real javascript error.

This is working good if you just print the stack to the console
however this is not much helpful for debugging as the printed
stacks is not clickable (Only Chrome prints clickable links but
it doesn't do that when a sourceUrl is provided and we need to
do that for superdev).

An alternative is logging the exception by re-throwing it via
timeout and let the console handle it. Even though this works
on Chrome (only when console is already open), it doesn't work
with custom error objects.

As the custom Error object concept still too futuristic for
today's browsers this patch goes back to more traditinal approach
by leting java Exception carry its backing javascript error with.
This is relativly simpler in implementation and gives ability for
loggers to print clickable stack by re-throwing the backing
exception or can let a compiler pass to unwrap it when thrown
like we do for JavaScriptException.

Change-Id: I43bc1c2b2deb74d863cfd8c59b281286f600c709