Fixes sporadic slow compiles by removing softMap caches.
Several caches were holding only "soft references" to cached objects so
that when under memory pressure the garbage collector could choose to 
clear some of the cached contents. It sounds reasonable but leads to 
hard-to-explain sporadically slow compiles after cache clears.

It seems better to hold onto all references so compiles are always
fast. If the compile runs out of memory it is probably best to throw an
OOM error so that the user will allocate more memory.

Change-Id: I14f180c8c42607c02dd06a1d01d7503adc7e8475
Review-Link: https://gwt-review.googlesource.com/#/c/9840/