Fixes a test to honor PathPrefixSet immutability.
PathPrefixSets are supposed to immutable in that they are built up 
during the ModuleDef parse process and then never modified.

Some PathPrefixSet usage is only safe as long as this assumption is 
honored, but there was a test that was not honoring the assumption.

Added documentation of the requirement and fixed the bad test.

Change-Id: I094c523eb700f075404954f353327d8b74a16804