Make cldr-import more deterministic
Wrap CldrFile in InputFile and Factory in InputFactory to ensure
that we don't depend on a HashMap's iteration order. This fixes
an issue where the 'TL' abbreviation for Turkish Lira wasn't
showing up in property files.

When writing out the list property files, sort the keys.

Stop including the date and type in the version file.
The date is in the local machine's timezone (due to Subversion).
The type is from whichever file was seen last.

Change-Id: I15fb375aeed08dded89dbed59f3ed09689cdaeef