Fixes handling of double mustache in UiBinder.
In UiBinder, when an attiribute is set to an expression with double mustache
the first curly bracket disappears.
 e.g. {{p in posts}} becomes {p in posts}}
This is related to some undocumented behavior to disable field reference
resoulution but it is pretty bad and  makes it diffucult to integrate UiBinder
with runtime templating systems like Polymer.

This patch introduces a  new escaping option: \{ABC} becomes {ABC} and not
treated as a field reference. In addition, double mustache is also not treated
as field reference nor replaced due escaping so that it will be correctly
preserved in the final output.

Change-Id: I24a202f3165a8ecdca8fb419e9c3197818df87b5