Allow compiler to emit all @defs of css files.
When using multiple .css files that declare variables in one
file and use them in another, Css2Gss currently fails because
of undefined variables.

This CL allows the GWT compiler to emit all @defs that were
declared inside of one compile and thus can be used with
Css2Gss to provide a proper scope for conversion.

Change-Id: Ifca4424f7705b9c8c249ea64633d52bd95c488a8