Apply optimizer-framework to more optimizers.
Now it has been applied to MakeCallsStatic, MethodInliner,
TypeTightener and DeadCodeElimination.

Tried applying it to MethodCallSpecializer and MethodCallTightener,
but had regression in compiling time. Both of them took actions on
a small part of the AST. So the overhead by applying the framework
exceeds the benefit.

Also tried removing the inner loop in TypeTightener, and it led to 2
more outerloops, which cost more time than the inner loop in
TypeTightener. Actually the inner loops (after the first iteration)
in TypeTightener costs little time. So just keep it.

The framework cannot be easily applied to Pruner and Finalizer, which
have to go through the whole AST.

Change-Id: Ib8143fdc00b15d1e8493f21904eb47534eb06ebc