Small fix in AnalizableTypeEnvironment.
As a result of performance tuning, the multimaps used in the
AnalizableTypeEnvironment might contain duplicates. In most cases
the removal is by key, so duplicates are removed all at once, but
in the case of overrides that was not the case.

The failure to remove some overrides is safe, resulting in a more
conservative prunning.

Change-Id: If5c5ce59789d2e224ba79e73e077bf5bdef28c91