Fixes ignored module property value changes in SDM.
Previously SDM was failing to invalidate its MinimalRebuildCache when 
property values were edited in .gwt.xml module files.

It did notice that the module files had been edited and treated it as a
recompile condition and it did retrieve/reuse a MinimalRebuildCache 
as keyed by the current set of binding properties. But that key 
construction was based on just the current set of binding properties as
being selected in the browser and ignored manual edits to them as 
performed in .gwt.xml file changes.

Change-Id: I952d42817abdc2000849f6b1f7b69de5b9420443