Fixes a bug that was exporting private constructors.
The access modifier for JConstructors was always being recorded as
public, which meant that private constructors used to not throw any
JsInterop validity check exceptions (which was good) but they were
being output (which was bad). With the new JsInteropRestrictionChecker
checking they started being reported as errors which is correct from
the perspective of blocking name collisions but is wrong from the
perspective that these methods shouldn't have been being exported
anyway.

Fixed the bug(s) by making sure that JConstructors are created with the
proper access level. As a result validity checks and output processes
are not running properly.

Change-Id: I561d227835238e4d8e714cd33e6a91b552b6eec7