Improve emul test slowness due to large char array initialization.
Tests are no longer reproducing the stack overflow issue because the
maximum stack size increased again for Chrome. We cannot keep enlarging
the array and also it will stop testing the issue silently when the
stack size increases again.
So instead now we choose a size that is smaller then earlier  but large
enough to test the batching logic.

Also
 - Improves the creation of sample String.
 - Removes the static initialization based init, as it is easy to
   accidentally slow down other tests.

Change-Id: Ia88de8be61e833b4e74acb9e226f844cc272016c
Review-Link: https://gwt-review.googlesource.com/#/c/12010/