Add restrictions to the exporting of ctors.
With this patch, now a constructor could only be exported either if it
is the only constructor in the class or if all other constructors are
delegating to it.

Also the exported can no longer supply a name. It will be always exported
by the qualified name of the type.

Exporting multiple constructors for the same object in different
namespace is problematic due to js instanceof checks and placing
static methods. Also in the long run, we are already moving towards
generating single constructor per class while generating user-defined
constructors as factory methods so this is making JsInterop inline
with that.

Change-Id: I4008afe6d86261c6de2c7bda3aa1dfad992123a4
Review-Link: https://gwt-review.googlesource.com/#/c/12634/