Fix sorting of place types in generated getTokenAndPrefix
The use of a TreeMap with a Comparator that falls back to type name for
unrelated types can lead to related types ending up in different branches,
and the generated getToken returning the wrong token (for a supertype,
when a more specific types exists).

Sorting by "flattened supertype hierarchy size" works here because:
 * Place is a class (rather than an interface),
 * the compared types are guaranteed to extend the Place type, and
 * the input to the getTokenAndPrefix is a Place too.

Original idea by larsaaslin@gmail.com.

Bug: issue 8086
Change-Id: I9295c9546839aa4cf87041c3f508acd575ffccfb