Fix AutoBeanFactory#equals in the JVM
factory.equals(factory) routed the equals to the InvocationHandler,
which would then compare an InvocationHandler to a proxy object using
the default Object#equals() implementation. Equals thus needs to be
overridden to extract the InvocationHandler out of the passed in object
when it is a proxy, before handing it to super.equals().

Change-Id: Ibd192631736751dc5dc4f536ed12865eae60a51d