Change the JsInterop bridge direction.
Earlier we were forwarding jsmethod at the first place it was exposed:

  parent.prototype.m = function() { return this.m_obfuscated() };

This caused a code regression in Closure based apps as Closure fail
to prune anything due to missing type information. This basically left
an alias bride method for every jsmethod.

I changed the implementation so that the direction is reversed.
So we generate methods with pretty js names and then create a bridge
from obfuscated name only if
 - method override a non-jsmethod
 - there is no parent who created such bridge.

Change-Id: I3ae2b023c8406abe42dcb2acc1a1271a873364e9