Update outdated links in samples
Showcase and DynaTable linked to the old doc and homepage at
code.google.com (not even developers.google.com!)

The goal in README-MAVEN.txt files is only to have working links, by
making them point to the backup of the old wiki on GitHub's
old_google_code_wiki repo.

Finally, replace the mention of GIN with Dagger in MobileWebApp's
ClientFactory.

Change-Id: I06ab002a68ecdf06bb791f13f615e293026ae766